% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_setup.R
\name{.check_dsOMOP_oracle_prereqs}
\alias{.check_dsOMOP_oracle_prereqs}
\title{Check dsOMOP.oracle prerequisites}
\usage{
.check_dsOMOP_oracle_prereqs()
}
\value{
A message indicating whether the required packages are installed.
}
\description{
This function checks if the required packages \code{dsOMOP} and \code{oracle.resourcer} are installed.
If any of the required packages are missing, it stops and returns an error message.
If both packages are installed, it returns a confirmation message.
}
\details{
Please note that this function is added to the package so that, when registered as a method,
the Opal UI correctly displays the \code{dsOMOP.oracle} package in the list of installed packages,
facilitating its management and discoverability by the data manager.
}
